[C++] Address some compiler warnings
See the recent commit history at https://bitbucket.org/jksh/lq-codegen/
for more details

+++ src/Cpp/lqtypes.cpp
@@ -8,7 +8,7 @@
 
 #include "lqtypes.h"
 
-static const int lStrHeaderSize = 4;
+static const size_t lStrHeaderSize = 4;
 
 void
 operator<<(LStrHandle dest, const QByteArray& src)
